Changed main package name
