+++ buildspec.yml
@@ -15,11 +15,11 @@ phases:
       - mkdir deploy/lib
       - cp target/scala-2.13/messaging_pipeline.jar deploy/lib
       - docker push aws_account_id.dkr.ecr.region.amazonaws.com/my-web-app
-      - java -classpath messaging_pipeline.jar me.skatz.producer.KafkaProducer
-      - java -classpath messaging_pipeline.jar me.skatz.LTProc.LTProc
-      - java -classpath messaging_pipeline.jar me.skatz.esProc.ElasticSearchProc
-      - java -classpath messaging_pipeline.jar me.skatz.enrichment.EnrichmentProc
-      - java -classpath messaging_pipeline.jar me.skatz.cassandraProc.CassandraProc
+      - java -classpath messaging_pipeline.jar me.messaging.producer.KafkaProducer
+      - java -classpath messaging_pipeline.jar me.messaging.LTProc.LTProc
+      - java -classpath messaging_pipeline.jar me.messaging.esProc.ElasticSearchProc
+      - java -classpath messaging_pipeline.jar me.messaging.enrichment.EnrichmentProc
+      - java -classpath messaging_pipeline.jar me.messaging.cassandraProc.CassandraProc
     artifacts:
       type: zip
       files:
