+++ src/main/scala/me/messaging/LTProc/metrics/Metrics.scala
@@ -1,0 +1,9 @@
+package me.messaging.LTProc.metrics
+
+import me.messaging.shared.metrics.Metric
+
+object Metrics {
+  object Metrics {
+    val MessagesSent = new Metric("LTProc.messages.processed", Metric.SUM)
+  }
+}
