+++ src/main/scala/me/messaging/cassandraProc/metrics/Metrics.scala
@@ -1,0 +1,8 @@
+package me.messaging.cassandraProc.metrics
+
+import me.messaging.shared.metrics.Metric
+
+object Metrics {
+  val MessagesSent = new Metric("cassandraProc.messages.processed", Metric.SUM)
+}
+
