+++ src/main/scala/me/messaging/enrichment/metrics/Metrics.scala
@@ -1,0 +1,7 @@
+package me.messaging.enrichment.metrics
+
+import me.messaging.shared.metrics.Metric
+
+object Metrics {
+  val MessagesSent = new Metric("enrichmentProc.messages.processed", Metric.SUM)
+}
