+++ src/main/scala/me/messaging/esProc/metrics/Metrics.scala
@@ -1,0 +1,7 @@
+package me.messaging.esProc.metrics
+
+import me.messaging.shared.metrics.Metric
+
+object Metrics {
+  val MessagesSent = new Metric("esProc.messages.processed", Metric.SUM)
+}
