+++ src/main/scala/me/messaging/producer/metrics/Metrics.scala
@@ -1,0 +1,7 @@
+package me.messaging.producer.metrics
+
+import me.messaging.shared.metrics.Metric
+
+object Metrics {
+  val MessagesSent: Metric = Metric("producer.messages.sent", Metric.SUM)
+}
