+++ src/main/scala/me/messaging/shared/JsonHelper.scala
@@ -1,0 +1,12 @@
+package me.messaging.shared
+
+import com.google.gson.Gson
+
+object JsonHelper {
+
+  def parseObject(obj: Any): String = {
+    val gson = new Gson
+    val jsonString = gson.toJson(obj)
+    jsonString
+  }
+}
