--- src/main/scala/me/skatz/LTProc/metrics/Metrics.scala
@@ -1,9 +1,0 @@
-package me.skatz.LTProc.metrics
-
-import me.skatz.shared.metrics.Metric
-
-object Metrics {
-  object Metrics {
-    val MessagesSent = new Metric("LTProc.messages.processed", Metric.SUM)
-  }
-}
