--- src/main/scala/me/skatz/cassandraProc/metrics/Metrics.scala
@@ -1,8 +1,0 @@
-package me.skatz.cassandraProc.metrics
-
-import me.skatz.shared.metrics.Metric
-
-object Metrics {
-  val MessagesSent = new Metric("cassandraProc.messages.processed", Metric.SUM)
-}
-
