--- src/main/scala/me/skatz/enrichment/metrics/Metrics.scala
@@ -1,7 +1,0 @@
-package me.skatz.enrichment.metrics
-
-import me.skatz.shared.metrics.Metric
-
-object Metrics {
-  val MessagesSent = new Metric("enrichmentProc.messages.processed", Metric.SUM)
-}
