--- src/main/scala/me/skatz/esProc/metrics/Metrics.scala
@@ -1,7 +1,0 @@
-package me.skatz.esProc.metrics
-
-import me.skatz.shared.metrics.Metric
-
-object Metrics {
-  val MessagesSent = new Metric("esProc.messages.processed", Metric.SUM)
-}
