--- src/main/scala/me/skatz/producer/metrics/Metrics.scala
@@ -1,7 +1,0 @@
-package me.skatz.producer.metrics
-
-import me.skatz.shared.metrics.Metric
-
-object Metrics {
-  val MessagesSent: Metric = Metric("producer.messages.sent", Metric.SUM)
-}
