--- src/main/scala/me/skatz/shared/Configuration.scala
@@ -1,36 +1,0 @@
-package me.skatz.shared
-
-object Configuration {
-
-  // Kafka
-  val bootstrapServer: String = sys.env.getOrElse("bootstrap_servers", "127.0.0.1:9092")
-  val autoOffsetReset: String = sys.env.getOrElse("auto_offset_reset", "latest")
-  val groupId: String = sys.env.getOrElse("group_id", "consumer-group")
-  val ingestEnrichTopic: String = sys.env.getOrElse("ingest_enrich_topic", "ingest_enrich")
-
-  // Elastic Search
-  val esUrl: String = sys.env.getOrElse("elasticsearch_url", "127.0.0.1")
-  val esPort: String = sys.env.getOrElse("elasticsearch_port", "9200")
-  val enrichEsprocTopic: String = sys.env.getOrElse("enrich_esproc_topic", "enrich_esproc")
-
-  // Cassandra
-  val cassandraUrl: String = sys.env.getOrElse("cassandra_url", "127.0.0.1")
-  val cassandraPort: String = sys.env.getOrElse("cassandra_port", "9042")
-  val keyspace: String = sys.env.getOrElse("cassandra_keyspace", "kafka")
-  val enrichCassTopic: String = sys.env.getOrElse("enrich_cassproc_topic", "enrich_cassproc")
-
-  // Metric Processor
-  val metricProcSourceTopic: String = sys.env.getOrElse("metricproc_source_topic", "metricproc_source")
-
-  // Long Time Processor
-  val s3Url: String = sys.env.getOrElse("s3_url", "http://127.0.0.1:9090")
-  val s3Region: String = sys.env.getOrElse("s3_region", "us-west-2")
-  val s3BucketName: String = sys.env.getOrElse("s3_bucket", "messaging")
-  val enrichLTTopic: String = sys.env.getOrElse("enrich_ltproc_topic", "enrich_ltproc")
-
-  // Files Directory
-  val filesDir: String = sys.env.getOrElse("files_dir", "/src/main/resources/twitter")
-  val firstNamesFile: String = sys.env.getOrElse("fname_file", "firstname.txt")
-  val lastNamesFile: String = sys.env.getOrElse("lname_file", "surname.txt")
-  val wordsFile: String = sys.env.getOrElse("words_file", "words.txt")
-}
