--- src/main/scala/me/skatz/shared/JsonHelper.scala
@@ -1,12 +1,0 @@
-package me.skatz.shared
-
-import com.google.gson.Gson
-
-object JsonHelper {
-
-  def parseObject(obj: Any): String = {
-    val gson = new Gson
-    val jsonString = gson.toJson(obj)
-    jsonString
-  }
-}
