--- src/main/scala/me/skatz/shared/metrics/Metric.scala
@@ -1,11 +1,0 @@
-package me.skatz.shared.metrics
-
-case class Metric(key: String, aggregator: Metric.Value)
-
-object Metric extends Enumeration {
-  type MetricValue = Value
-  val MIN = Value("MIN")
-  val MAX = Value("MAX")
-  val SUM = Value("SUM")
-}
-
